Step based DQN

+++ deep_rl/utils/torch_utils.py
@@ -22,7 +22,7 @@ def tensor(x):
     return x
 
 def range_tensor(end):
-    return torch.arange(end).to(Config.DEVICE)
+    return torch.arange(end).long().to(Config.DEVICE)
 
 def to_np(t):
     return t.cpu().detach().numpy()
